api/requirements.txt
alembic==1.12.1
algoliasearch==3.0.0
bcrypt==4.0.1
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
